Full 64 bit indexing support throughout Caffe.

+++ src/caffe/util/math_functions.cpp
@@ -262,7 +262,7 @@ double caffe_nextafter(const double b);
 void caffe_rng_uniform(const int_tp n, uint_tp* r) {
   CHECK_GE(n, 0);
   CHECK(r);
-  boost::uniform_int<int_tp> random_distribution(INT32_MIN, INT32_MAX);
+  boost::uniform_int<int_tp> random_distribution(INT64_MIN, INT64_MAX);
   boost::variate_generator<caffe::rng_t*, boost::uniform_int<int_tp>>
     variate_generator(caffe_rng(), random_distribution);
   for (int_tp i = 0; i < n; ++i) {
